+++ embulk-output-sqlserver/src/test/java/org/embulk/output/sqlserver/SQLServerOutputPluginTest.java
@@ -2,32 +2,21 @@ package org.embulk.output.sqlserver;
 
 import static org.junit.Assert.assertEquals;
 
-import java.io.File;
-import java.net.URISyntaxException;
-import java.nio.charset.Charset;
 import java.sql.Connection;
 import java.sql.DriverManager;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
 import java.util.TimeZone;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
+import org.embulk.output.AbstractJdbcOutputPluginTest;
 import org.embulk.output.SQLServerOutputPlugin;
 import org.embulk.output.tester.EmbulkPluginTester;
 import org.embulk.spi.OutputPlugin;
 import org.junit.Test;
 
-import com.google.common.io.Files;
-
-public class SQLServerOutputPluginTest {
-
+public class SQLServerOutputPluginTest extends AbstractJdbcOutputPluginTest
+{
     private static EmbulkPluginTester tester = new EmbulkPluginTester();
     static {
         tester.addPlugin(OutputPlugin.class, "sqlserver", SQLServerOutputPlugin.class);
@@ -88,12 +77,6 @@ public class SQLServerOutputPluginTest {
         }
     }
 
-    private void dropTable(String table) throws SQLException
-    {
-        String sql = String.format("DROP TABLE %s", table);
-        executeSQL(sql, true);
-    }
-
     private void createTable(String table) throws SQLException
     {
         String sql = String.format("CREATE TABLE %s ("
@@ -111,88 +94,8 @@ public class SQLServerOutputPluginTest {
         executeSQL(String.format("INSERT INTO %s VALUES('9999', NULL, NULL, NULL, NULL)", table));
     }
 
-    private List<List<Object>> select(String table) throws SQLException
-    {
-        try (Connection connection = connect()) {
-            try (Statement statement = connection.createStatement()) {
-                List<List<Object>> rows = new ArrayList<List<Object>>();
-                String sql = String.format("SELECT * FROM %s ORDER BY ID", table);
-                System.out.println(sql);
-                try (ResultSet resultSet = statement.executeQuery(sql)) {
-                    while (resultSet.next()) {
-                        List<Object> row = new ArrayList<Object>();
-                        for (int i = 1; i <= resultSet.getMetaData().getColumnCount(); i++) {
-                            Object value = resultSet.getObject(i);
-                            if (value != null && value.getClass().getName().equals("oracle.sql.CLOB")) {
-                                value = resultSet.getString(i);
-                            }
-                            row.add(value);
-                        }
-                        rows.add(row);
-                    }
-                }
-                // cannot sort by CLOB, so sort by Java
-                Collections.sort(rows, new Comparator<List<Object>>() {
-                    @Override
-                    public int compare(List<Object> o1, List<Object> o2) {
-                        return o1.toString().compareTo(o2.toString());
-                    }
-                });
-                return rows;
-            }
-        }
-
-    }
-
-    private void executeSQL(String sql) throws SQLException
-    {
-        executeSQL(sql, false);
-    }
-
-    private void executeSQL(String sql, boolean ignoreError) throws SQLException
-    {
-        try (Connection connection = connect()) {
-            try {
-                connection.setAutoCommit(true);
-
-                try (Statement statement = connection.createStatement()) {
-                    System.out.println(String.format("Execute SQL : \"%s\".", sql));
-                    statement.execute(sql);
-                }
-
-            } catch (SQLException e) {
-                if (!ignoreError) {
-                    throw e;
-                }
-            }
-        }
-    }
-
-    private String convertYml(String ymlName) throws Exception
-    {
-        StringBuilder builder = new StringBuilder();
-        Pattern pathPrefixPattern = Pattern.compile("^ *path_prefix: '(.*)'$");
-        for (String line : Files.readLines(convertPath(ymlName), Charset.defaultCharset())) {
-            Matcher matcher = pathPrefixPattern.matcher(line);
-            if (matcher.matches()) {
-                int group = 1;
-                builder.append(line.substring(0, matcher.start(group)));
-                builder.append(convertPath(matcher.group(group)).getAbsolutePath());
-                builder.append(line.substring(matcher.end(group)));
-            } else {
-                builder.append(line);
-            }
-            builder.append(System.lineSeparator());
-        }
-        return builder.toString();
-    }
-
-    private File convertPath(String name) throws URISyntaxException
-    {
-        return new File(getClass().getResource(name).toURI());
-    }
-
-    private static Connection connect() throws SQLException
+    @Override
+    protected Connection connect() throws SQLException
     {
         return DriverManager.getConnection("jdbc:sqlserver://localhost\\SQLEXPRESS:1433;databasename=TESTDB", "TEST_USER", "test_pw");
     }
